+++ extensions/dashlane-vault/src/lib/dcli.ts
@@ -2,7 +2,7 @@ import { getPreferenceValues } from "@raycast/api";
 import { existsSync } from "fs";
 import { safeParse } from "valibot";
 
-import { execFilePromis } from "@/helper/exec";
+import { execFilePromise } from "@/helper/exec";
 import { VaultCredential, VaultCredentialSchema, VaultNote, VaultNoteSchema } from "@/types/dcli";
 
 const preferences = getPreferenceValues<Preferences>();
@@ -11,12 +11,13 @@ const CLI_PATH =
   preferences.cliPath ?? ["/usr/local/bin/dcli", "/opt/homebrew/bin/dcli"].find((path) => existsSync(path));
 
 async function dcli(...args: string[]) {
-  if (CLI_PATH) {
-    const { stdout } = await execFilePromis(CLI_PATH, args, { maxBuffer: 4096 * 1024 });
-    return stdout;
+  if (!CLI_PATH) {
+    throw Error("Dashlane CLI is not found!");
   }
 
-  throw Error("Dashlane CLI is not found!");
+  const { stdout } = await execFilePromise(CLI_PATH, args, { maxBuffer: 4096 * 1024 });
+
+  return stdout;
 }
 
 export async function syncVault() {
@@ -42,13 +43,17 @@ export async function getNotes() {
 }
 
 export async function getPassword(id: string) {
-  const stdout = await dcli("password", `id=${id}`, "--output", "password");
+  const stdout = await dcli("read", `dl://${extractId(id)}/password`);
   return stdout.trim();
 }
 
 export async function getOtpSecret(id: string) {
-  const stdout = await dcli("otp", `id=${id}`, "--print");
-  return stdout.trim();
+  const result = await dcli("read", `dl://${extractId(id)}/otpSecret?otp+expiry`);
+  const [otp, expireIn] = result.split(" ").map((item) => item.trim());
+  return {
+    otp,
+    expireIn,
+  };
 }
 
 function parseVaultCredentials(jsonString: string): VaultCredential[] {
@@ -89,4 +94,15 @@ function parseNotes(jsonString: string): VaultNote[] {
   } catch (error) {
     return [];
   }
+}
+
+/**
+ * Dashlane CLI returns the ID in the format of `{id}`.
+ * @returns Id without curly braces.
+ */
+function extractId(id: string) {
+  if (id.startsWith("{") && id.endsWith("}")) {
+    return id.slice(1, -1);
+  }
+  return id;
 }